Fixed the rectangle collisions
I have removed the "remove" option in the collision method
since it is horribly complicated enough as it is. I might
implement it later as it probably a common enough case... (not terribly
clear if it would be worth it, as it will incur some overhead
for other methods not concerned with removal).

It may even be better to hardcode the arrays, but this much less
memory efficient.